Generate the PMML representation for a cv.glmnet object from the package glmnet.

Generate the PMML representation for a cv.glmnet object from the package glmnet.

# S3 method for cv.glmnet
pmml(model, model_name = "Elasticnet_Model",
  app_name = "SoftwareAG PMML Generator",
  description = "Generalized Linear Regression Model",
  copyright = NULL, transforms = NULL,
  missing_value_replacement = NULL, dataset = NULL, s = NULL, ...)

Arguments

model
A cv.glmnet object.

model_name
A name to be given to the PMML model.

app_name
The name of the application that generated the PMML.

description
A descriptive text for the Header element of the PMML.

copyright
The copyright notice for the model.

transforms
Data transformations.

missing_value_replacement
Value to be used as the 'missingValueReplacement' attribute for all MiningFields.

dataset
Data used to train the cv.glmnet model.

s
'lambda' parameter at which to output the model. If not given, the lambda.1se parameter from the model is used instead.

...
Further arguments passed to or from other methods.

Value

PMML representation of the cv.glmnet object.

Details

The glmnet package expects the input and predicted values in a matrix format - not as arrays or data frames. As of now, it will also accept numerical values only. As such, any string variables must be converted to numerical ones. One possible way to do so is to use data transformation functions from this package. However, the result is a data frame. In all cases, lists, arrays and data frames can be converted to a matrix format using the data.matrix function from the base package. Given a data frame df, a matrix m can thus be created by using m <- data.matrix(df).

The PMML language requires variable names which will be read in as the column names of the input matrix. If the matrix does not have variable names, they will be given the default values of "X1", "X2", ...

Currently, only gaussian and poisson family types are supported.

References

glmnet: Lasso and elastic-net regularized generalized linear models (on CRAN)

Examples

library(glmnet)

#> Loading required package: Matrix

#> Loading required package: foreach

#> Loaded glmnet 2.0-18

# Create a simple predictor (x) and response(y) matrices:
x <- matrix(rnorm(100 * 20), 100, 20)
y <- rnorm(100)

# Build a simple gaussian model:
model1 <- cv.glmnet(x, y)

# Output the model in PMML format:
pmml(model1)

#> <PMML version="4.3" xmlns="http://www.dmg.org/PMML-4_3" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.dmg.org/PMML-4_3 http://www.dmg.org/pmml/v4-3/pmml-4-3.xsd">
#>  <Header copyright="Copyright (c) 2019 dmitriy" description="Generalized Linear Regression Model">
#>   <Extension name="user" value="dmitriy" extender="SoftwareAG PMML Generator"/>
#>   <Application name="SoftwareAG PMML Generator" version="1.4"/>
#>   <Timestamp>2019-05-30 22:53:58</Timestamp>
#>  </Header>
#>  <DataDictionary numberOfFields="21">
#>   <DataField name="predictedScore" optype="continuous" dataType="double"/>
#>   <DataField name="V1" optype="continuous" dataType="double"/>
#>   <DataField name="V2" optype="continuous" dataType="double"/>
#>   <DataField name="V3" optype="continuous" dataType="double"/>
#>   <DataField name="V4" optype="continuous" dataType="double"/>
#>   <DataField name="V5" optype="continuous" dataType="double"/>
#>   <DataField name="V6" optype="continuous" dataType="double"/>
#>   <DataField name="V7" optype="continuous" dataType="double"/>
#>   <DataField name="V8" optype="continuous" dataType="double"/>
#>   <DataField name="V9" optype="continuous" dataType="double"/>
#>   <DataField name="V10" optype="continuous" dataType="double"/>
#>   <DataField name="V11" optype="continuous" dataType="double"/>
#>   <DataField name="V12" optype="continuous" dataType="double"/>
#>   <DataField name="V13" optype="continuous" dataType="double"/>
#>   <DataField name="V14" optype="continuous" dataType="double"/>
#>   <DataField name="V15" optype="continuous" dataType="double"/>
#>   <DataField name="V16" optype="continuous" dataType="double"/>
#>   <DataField name="V17" optype="continuous" dataType="double"/>
#>   <DataField name="V18" optype="continuous" dataType="double"/>
#>   <DataField name="V19" optype="continuous" dataType="double"/>
#>   <DataField name="V20" optype="continuous" dataType="double"/>
#>  </DataDictionary>
#>  <GeneralRegressionModel modelName="Elasticnet_Model" modelType="generalLinear" algorithmName="glmnet" functionName="regression">
#>   <Extension name="lambda" value="0.188441538440095"/>
#>   <MiningSchema>
#>    <MiningField name="predictedScore" usageType="predicted" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="V1" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="V2" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="V3" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="V4" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="V5" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="V6" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="V7" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="V8" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="V9" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="V10" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="V11" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="V12" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="V13" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="V14" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="V15" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="V16" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="V17" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="V18" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="V19" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="V20" usageType="active" invalidValueTreatment="returnInvalid"/>
#>   </MiningSchema>
#>   <Output>
#>    <OutputField name="predictedValue" feature="predictedValue" dataType="double" optype="continuous"/>
#>   </Output>
#>   <ParameterList>
#>    <Parameter name="p0" label="Intercept"/>
#>    <Parameter name="p1" label="V1"/>
#>    <Parameter name="p2" label="V2"/>
#>    <Parameter name="p3" label="V3"/>
#>    <Parameter name="p4" label="V4"/>
#>    <Parameter name="p5" label="V5"/>
#>    <Parameter name="p6" label="V6"/>
#>    <Parameter name="p7" label="V7"/>
#>    <Parameter name="p8" label="V8"/>
#>    <Parameter name="p9" label="V9"/>
#>    <Parameter name="p10" label="V10"/>
#>    <Parameter name="p11" label="V11"/>
#>    <Parameter name="p12" label="V12"/>
#>    <Parameter name="p13" label="V13"/>
#>    <Parameter name="p14" label="V14"/>
#>    <Parameter name="p15" label="V15"/>
#>    <Parameter name="p16" label="V16"/>
#>    <Parameter name="p17" label="V17"/>
#>    <Parameter name="p18" label="V18"/>
#>    <Parameter name="p19" label="V19"/>
#>    <Parameter name="p20" label="V20"/>
#>   </ParameterList>
#>   <CovariateList>
#>    <Predictor name="V1"/>
#>    <Predictor name="V2"/>
#>    <Predictor name="V3"/>
#>    <Predictor name="V4"/>
#>    <Predictor name="V5"/>
#>    <Predictor name="V6"/>
#>    <Predictor name="V7"/>
#>    <Predictor name="V8"/>
#>    <Predictor name="V9"/>
#>    <Predictor name="V10"/>
#>    <Predictor name="V11"/>
#>    <Predictor name="V12"/>
#>    <Predictor name="V13"/>
#>    <Predictor name="V14"/>
#>    <Predictor name="V15"/>
#>    <Predictor name="V16"/>
#>    <Predictor name="V17"/>
#>    <Predictor name="V18"/>
#>    <Predictor name="V19"/>
#>    <Predictor name="V20"/>
#>   </CovariateList>
#>   <PPMatrix>
#>    <PPCell value="1" predictorName="V1" parameterName="p1"/>
#>    <PPCell value="1" predictorName="V2" parameterName="p2"/>
#>    <PPCell value="1" predictorName="V3" parameterName="p3"/>
#>    <PPCell value="1" predictorName="V4" parameterName="p4"/>
#>    <PPCell value="1" predictorName="V5" parameterName="p5"/>
#>    <PPCell value="1" predictorName="V6" parameterName="p6"/>
#>    <PPCell value="1" predictorName="V7" parameterName="p7"/>
#>    <PPCell value="1" predictorName="V8" parameterName="p8"/>
#>    <PPCell value="1" predictorName="V9" parameterName="p9"/>
#>    <PPCell value="1" predictorName="V10" parameterName="p10"/>
#>    <PPCell value="1" predictorName="V11" parameterName="p11"/>
#>    <PPCell value="1" predictorName="V12" parameterName="p12"/>
#>    <PPCell value="1" predictorName="V13" parameterName="p13"/>
#>    <PPCell value="1" predictorName="V14" parameterName="p14"/>
#>    <PPCell value="1" predictorName="V15" parameterName="p15"/>
#>    <PPCell value="1" predictorName="V16" parameterName="p16"/>
#>    <PPCell value="1" predictorName="V17" parameterName="p17"/>
#>    <PPCell value="1" predictorName="V18" parameterName="p18"/>
#>    <PPCell value="1" predictorName="V19" parameterName="p19"/>
#>    <PPCell value="1" predictorName="V20" parameterName="p20"/>
#>   </PPMatrix>
#>   <ParamMatrix>
#>    <PCell parameterName="p0" df="1" beta="0.0949216513133108"/>
#>   </ParamMatrix>
#>  </GeneralRegressionModel>
#> </PMML>

# Shift y between 0 and 1 to create a poisson response:
y <- y - min(y)

# Give the predictor variables names (default values are V1,V2,...):
name <- NULL
for (i in 1:20) {
  name <- c(name, paste("variable", i, sep = ""))
}
colnames(x) <- name

# Create a simple poisson model:
model2 <- cv.glmnet(x, y, family = "poisson")

# Output the regression model in PMML format at the lambda
# parameter = 0.006:
pmml(model2, s = 0.006)

#> <PMML version="4.3" xmlns="http://www.dmg.org/PMML-4_3" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.dmg.org/PMML-4_3 http://www.dmg.org/pmml/v4-3/pmml-4-3.xsd">
#>  <Header copyright="Copyright (c) 2019 dmitriy" description="Generalized Linear Regression Model">
#>   <Extension name="user" value="dmitriy" extender="SoftwareAG PMML Generator"/>
#>   <Application name="SoftwareAG PMML Generator" version="1.4"/>
#>   <Timestamp>2019-05-30 22:53:59</Timestamp>
#>  </Header>
#>  <DataDictionary numberOfFields="21">
#>   <DataField name="predictedScore" optype="continuous" dataType="double"/>
#>   <DataField name="variable1" optype="continuous" dataType="double"/>
#>   <DataField name="variable2" optype="continuous" dataType="double"/>
#>   <DataField name="variable3" optype="continuous" dataType="double"/>
#>   <DataField name="variable4" optype="continuous" dataType="double"/>
#>   <DataField name="variable5" optype="continuous" dataType="double"/>
#>   <DataField name="variable6" optype="continuous" dataType="double"/>
#>   <DataField name="variable7" optype="continuous" dataType="double"/>
#>   <DataField name="variable8" optype="continuous" dataType="double"/>
#>   <DataField name="variable9" optype="continuous" dataType="double"/>
#>   <DataField name="variable10" optype="continuous" dataType="double"/>
#>   <DataField name="variable11" optype="continuous" dataType="double"/>
#>   <DataField name="variable12" optype="continuous" dataType="double"/>
#>   <DataField name="variable13" optype="continuous" dataType="double"/>
#>   <DataField name="variable14" optype="continuous" dataType="double"/>
#>   <DataField name="variable15" optype="continuous" dataType="double"/>
#>   <DataField name="variable16" optype="continuous" dataType="double"/>
#>   <DataField name="variable17" optype="continuous" dataType="double"/>
#>   <DataField name="variable18" optype="continuous" dataType="double"/>
#>   <DataField name="variable19" optype="continuous" dataType="double"/>
#>   <DataField name="variable20" optype="continuous" dataType="double"/>
#>  </DataDictionary>
#>  <GeneralRegressionModel modelName="Elasticnet_Model" modelType="generalLinear" algorithmName="glmnet" functionName="regression">
#>   <Extension name="lambda" value="0.006"/>
#>   <MiningSchema>
#>    <MiningField name="predictedScore" usageType="predicted" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="variable1" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="variable2" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="variable3" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="variable4" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="variable5" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="variable6" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="variable7" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="variable8" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="variable9" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="variable10" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="variable11" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="variable12" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="variable13" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="variable14" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="variable15" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="variable16" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="variable17" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="variable18" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="variable19" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="variable20" usageType="active" invalidValueTreatment="returnInvalid"/>
#>   </MiningSchema>
#>   <Output>
#>    <OutputField name="predictedValue" feature="predictedValue" dataType="double" optype="continuous"/>
#>    <OutputField name="predictedMean" feature="transformedValue" dataType="double">
#>     <Apply function="exp">
#>      <FieldRef field="predictedValue"/>
#>     </Apply>
#>    </OutputField>
#>   </Output>
#>   <ParameterList>
#>    <Parameter name="p0" label="Intercept"/>
#>    <Parameter name="p1" label="variable1"/>
#>    <Parameter name="p2" label="variable2"/>
#>    <Parameter name="p3" label="variable3"/>
#>    <Parameter name="p4" label="variable4"/>
#>    <Parameter name="p5" label="variable5"/>
#>    <Parameter name="p6" label="variable6"/>
#>    <Parameter name="p7" label="variable7"/>
#>    <Parameter name="p8" label="variable8"/>
#>    <Parameter name="p9" label="variable9"/>
#>    <Parameter name="p10" label="variable10"/>
#>    <Parameter name="p11" label="variable11"/>
#>    <Parameter name="p12" label="variable12"/>
#>    <Parameter name="p13" label="variable13"/>
#>    <Parameter name="p14" label="variable14"/>
#>    <Parameter name="p15" label="variable15"/>
#>    <Parameter name="p16" label="variable16"/>
#>    <Parameter name="p17" label="variable17"/>
#>    <Parameter name="p18" label="variable18"/>
#>    <Parameter name="p19" label="variable19"/>
#>    <Parameter name="p20" label="variable20"/>
#>   </ParameterList>
#>   <CovariateList>
#>    <Predictor name="variable1"/>
#>    <Predictor name="variable2"/>
#>    <Predictor name="variable3"/>
#>    <Predictor name="variable4"/>
#>    <Predictor name="variable5"/>
#>    <Predictor name="variable6"/>
#>    <Predictor name="variable7"/>
#>    <Predictor name="variable8"/>
#>    <Predictor name="variable9"/>
#>    <Predictor name="variable10"/>
#>    <Predictor name="variable11"/>
#>    <Predictor name="variable12"/>
#>    <Predictor name="variable13"/>
#>    <Predictor name="variable14"/>
#>    <Predictor name="variable15"/>
#>    <Predictor name="variable16"/>
#>    <Predictor name="variable17"/>
#>    <Predictor name="variable18"/>
#>    <Predictor name="variable19"/>
#>    <Predictor name="variable20"/>
#>   </CovariateList>
#>   <PPMatrix>
#>    <PPCell value="1" predictorName="variable1" parameterName="p1"/>
#>    <PPCell value="1" predictorName="variable2" parameterName="p2"/>
#>    <PPCell value="1" predictorName="variable3" parameterName="p3"/>
#>    <PPCell value="1" predictorName="variable4" parameterName="p4"/>
#>    <PPCell value="1" predictorName="variable5" parameterName="p5"/>
#>    <PPCell value="1" predictorName="variable6" parameterName="p6"/>
#>    <PPCell value="1" predictorName="variable7" parameterName="p7"/>
#>    <PPCell value="1" predictorName="variable8" parameterName="p8"/>
#>    <PPCell value="1" predictorName="variable9" parameterName="p9"/>
#>    <PPCell value="1" predictorName="variable10" parameterName="p10"/>
#>    <PPCell value="1" predictorName="variable11" parameterName="p11"/>
#>    <PPCell value="1" predictorName="variable12" parameterName="p12"/>
#>    <PPCell value="1" predictorName="variable13" parameterName="p13"/>
#>    <PPCell value="1" predictorName="variable14" parameterName="p14"/>
#>    <PPCell value="1" predictorName="variable15" parameterName="p15"/>
#>    <PPCell value="1" predictorName="variable16" parameterName="p16"/>
#>    <PPCell value="1" predictorName="variable17" parameterName="p17"/>
#>    <PPCell value="1" predictorName="variable18" parameterName="p18"/>
#>    <PPCell value="1" predictorName="variable19" parameterName="p19"/>
#>    <PPCell value="1" predictorName="variable20" parameterName="p20"/>
#>   </PPMatrix>
#>   <ParamMatrix>
#>    <PCell parameterName="p0" df="1" beta="0.792986102954512"/>
#>    <PCell parameterName="p1" df="1" beta="-0.0285965996765635"/>
#>    <PCell parameterName="p2" df="1" beta="0.0634169521465928"/>
#>    <PCell parameterName="p5" df="1" beta="-0.0261366055683498"/>
#>    <PCell parameterName="p6" df="1" beta="-0.00731136035858968"/>
#>    <PCell parameterName="p7" df="1" beta="-0.027374239591406"/>
#>    <PCell parameterName="p8" df="1" beta="0.0539945807369908"/>
#>    <PCell parameterName="p9" df="1" beta="-0.00805236058694793"/>
#>    <PCell parameterName="p10" df="1" beta="-0.000122789571110577"/>
#>    <PCell parameterName="p11" df="1" beta="-0.0150758280719763"/>
#>    <PCell parameterName="p12" df="1" beta="-0.0624100541192502"/>
#>    <PCell parameterName="p13" df="1" beta="-0.0176972849566792"/>
#>    <PCell parameterName="p14" df="1" beta="0.0230632410394262"/>
#>    <PCell parameterName="p15" df="1" beta="0.00429379572557985"/>
#>    <PCell parameterName="p16" df="1" beta="0.0832877681038215"/>
#>    <PCell parameterName="p17" df="1" beta="0.0162814778822209"/>
#>    <PCell parameterName="p18" df="1" beta="-0.00853386425279652"/>
#>    <PCell parameterName="p19" df="1" beta="0.0142102068710642"/>
#>    <PCell parameterName="p20" df="1" beta="0.0392343055678505"/>
#>   </ParamMatrix>
#>  </GeneralRegressionModel>
#> </PMML>