Generate the PMML representation for a gbm object from the package gbm.

Generate the PMML representation for a gbm object from the package gbm.

# S3 method for gbm
pmml(model, model_name = "GBM_Model",
  app_name = "SoftwareAG PMML Generator",
  description = "Generalized Boosted Tree Model", copyright = NULL,
  transforms = NULL, missing_value_replacement = NULL, ...)

Arguments

model
A gbm object.

model_name
A name to be given to the PMML model.

app_name
The name of the application that generated the PMML.

description
A descriptive text for the Header element of the PMML.

copyright
The copyright notice for the model.

transforms
Data transformations.

missing_value_replacement
Value to be used as the 'missingValueReplacement' attribute for all MiningFields.

...
Further arguments passed to or from other methods.

Value

PMML representation of the gbm object.

Details

The 'gbm' function uses various distribution types to fit a model; currently only the "bernoulli", "poisson" and "multinomial" distribution types are supported.

For all cases, the model output includes the gbm prediction type "link" and "response".

References

gbm: Generalized Boosted Regression Models (on CRAN)

Examples

library(gbm)

#> Loaded gbm 2.1.5

data(audit)

mod <- gbm(Adjusted ~ .,
  data = audit[, -c(1, 4, 6, 9, 10, 11, 12)],
  n.trees = 3, interaction.depth = 4
)

#> Distribution not specified, assuming bernoulli ...

pmml(mod)

#> Now converting tree  1  to PMML, for category# 1

#> Now converting tree  2  to PMML, for category# 1

#> Now converting tree  3  to PMML, for category# 1

#> <PMML version="4.3" xmlns="http://www.dmg.org/PMML-4_3" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.dmg.org/PMML-4_3 http://www.dmg.org/pmml/v4-3/pmml-4-3.xsd">
#>  <Header copyright="Copyright (c) 2019 dmitriy" description="Generalized Boosted Tree Model">
#>   <Extension name="user" value="dmitriy" extender="SoftwareAG PMML Generator"/>
#>   <Application name="SoftwareAG PMML Generator" version="1.4"/>
#>   <Timestamp>2019-06-04 17:04:22</Timestamp>
#>  </Header>
#>  <DataDictionary numberOfFields="6">
#>   <DataField name="Adjusted" optype="continuous" dataType="double"/>
#>   <DataField name="Age" optype="continuous" dataType="double"/>
#>   <DataField name="Employment" optype="categorical" dataType="string">
#>    <Value value="Consultant"/>
#>    <Value value="Private"/>
#>    <Value value="PSFederal"/>
#>    <Value value="PSLocal"/>
#>    <Value value="PSState"/>
#>    <Value value="SelfEmp"/>
#>    <Value value="Volunteer"/>
#>   </DataField>
#>   <DataField name="Marital" optype="categorical" dataType="string">
#>    <Value value="Absent"/>
#>    <Value value="Divorced"/>
#>    <Value value="Married"/>
#>    <Value value="Married-spouse-absent"/>
#>    <Value value="Unmarried"/>
#>    <Value value="Widowed"/>
#>   </DataField>
#>   <DataField name="Income" optype="continuous" dataType="double"/>
#>   <DataField name="Sex" optype="categorical" dataType="string">
#>    <Value value="Female"/>
#>    <Value value="Male"/>
#>   </DataField>
#>  </DataDictionary>
#>  <MiningModel modelName="GBM_Model" functionName="regression">
#>   <MiningSchema>
#>    <MiningField name="Adjusted" usageType="predicted" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="Age" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="Employment" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="Marital" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="Income" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="Sex" usageType="active" invalidValueTreatment="returnInvalid"/>
#>   </MiningSchema>
#>   <Output>
#>    <OutputField name="Predicted_Adjusted" optype="continuous" dataType="double" feature="predictedValue"/>
#>    <OutputField name="BernoulliLinkPrediction" feature="transformedValue" dataType="double">
#>     <Apply function="+">
#>      <FieldRef field="Predicted_Adjusted"/>
#>      <Constant>-1.56645402987148</Constant>
#>     </Apply>
#>    </OutputField>
#>    <OutputField name="BernoulliResponsePrediction" feature="transformedValue" dataType="double">
#>     <Apply function="/">
#>      <Constant>1</Constant>
#>      <Apply function="+">
#>       <Constant>1</Constant>
#>       <Apply function="exp">
#>        <Apply function="*">
#>         <Constant>-1</Constant>
#>         <FieldRef field="BernoulliLinkPrediction"/>
#>        </Apply>
#>       </Apply>
#>      </Apply>
#>     </Apply>
#>    </OutputField>
#>   </Output>
#>   <Segmentation multipleModelMethod="sum">
#>    <Segment id="1">
#>     <True/>
#>     <TreeModel modelName="GBM_Model" functionName="regression" algorithmName="gbm" splitCharacteristic="multiSplit">
#>      <MiningSchema>
#>       <MiningField name="Adjusted" usageType="predicted" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Age" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Employment" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Marital" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Income" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Sex" usageType="active" invalidValueTreatment="returnInvalid"/>
#>      </MiningSchema>
#>      <Node id="1">
#>       <True/>
#>       <Node id="2" score="-0.114313300974689">
#>        <SimpleSetPredicate field="Marital" booleanOperator="isIn">
#>         <Array n="5" type="string">&quot;Absent&quot;   &quot;Divorced&quot;   &quot;Married-spouse-absent&quot;   &quot;Unmarried&quot;   &quot;Widowed&quot;</Array>
#>        </SimpleSetPredicate>
#>       </Node>
#>       <Node id="3" score="0.131468394485593">
#>        <SimpleSetPredicate field="Marital" booleanOperator="isIn">
#>         <Array n="1" type="string">&quot;Married&quot;</Array>
#>        </SimpleSetPredicate>
#>        <Node id="4" score="-0.0593538643747457">
#>         <SimplePredicate field="Age" operator="lessThan" value="32.5"/>
#>        </Node>
#>        <Node id="5" score="0.184898626966487">
#>         <SimplePredicate field="Age" operator="greaterOrEqual" value="32.5"/>
#>         <Node id="6" score="0.394794510417884">
#>          <SimplePredicate field="Income" operator="lessThan" value="11162.46"/>
#>         </Node>
#>         <Node id="7" score="0.171865876033231">
#>          <SimplePredicate field="Income" operator="greaterOrEqual" value="11162.46"/>
#>          <Node id="8" score="0.149763538636402">
#>           <SimpleSetPredicate field="Employment" booleanOperator="isIn">
#>            <Array n="3" type="string">&quot;Consultant&quot;   &quot;Private&quot;   &quot;PSFederal&quot;</Array>
#>           </SimpleSetPredicate>
#>          </Node>
#>          <Node id="9" score="0.285029843504992">
#>           <SimpleSetPredicate field="Employment" booleanOperator="isIn">
#>            <Array n="4" type="string">&quot;PSLocal&quot;   &quot;PSState&quot;   &quot;SelfEmp&quot;   &quot;Volunteer&quot;</Array>
#>           </SimpleSetPredicate>
#>          </Node>
#>          <Node id="10" score="0.171865876033231">
#>           <SimplePredicate field="Employment" operator="isMissing"/>
#>          </Node>
#>         </Node>
#>         <Node id="11" score="0.184898626966487">
#>          <SimplePredicate field="Income" operator="isMissing"/>
#>         </Node>
#>        </Node>
#>        <Node id="12" score="0.131468394485593">
#>         <SimplePredicate field="Age" operator="isMissing"/>
#>        </Node>
#>       </Node>
#>       <Node id="13" score="-0.00657337967703121">
#>        <SimplePredicate field="Marital" operator="isMissing"/>
#>       </Node>
#>      </Node>
#>     </TreeModel>
#>    </Segment>
#>    <Segment id="2">
#>     <True/>
#>     <TreeModel modelName="GBM_Model" functionName="regression" algorithmName="gbm" splitCharacteristic="multiSplit">
#>      <MiningSchema>
#>       <MiningField name="Adjusted" usageType="predicted" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Age" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Employment" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Marital" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Income" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Sex" usageType="active" invalidValueTreatment="returnInvalid"/>
#>      </MiningSchema>
#>      <Node id="1">
#>       <True/>
#>       <Node id="2" score="-0.114195813526704">
#>        <SimpleSetPredicate field="Marital" booleanOperator="isIn">
#>         <Array n="5" type="string">&quot;Absent&quot;   &quot;Divorced&quot;   &quot;Married-spouse-absent&quot;   &quot;Unmarried&quot;   &quot;Widowed&quot;</Array>
#>        </SimpleSetPredicate>
#>       </Node>
#>       <Node id="3" score="0.110626908541706">
#>        <SimpleSetPredicate field="Marital" booleanOperator="isIn">
#>         <Array n="1" type="string">&quot;Married&quot;</Array>
#>        </SimpleSetPredicate>
#>        <Node id="4" score="0.0304204715377185">
#>         <SimplePredicate field="Age" operator="lessThan" value="40.5"/>
#>         <Node id="5" score="-0.0486579370164779">
#>          <SimplePredicate field="Age" operator="lessThan" value="31.5"/>
#>         </Node>
#>         <Node id="6" score="0.0844573840497526">
#>          <SimplePredicate field="Age" operator="greaterOrEqual" value="31.5"/>
#>          <Node id="7" score="0.0115790366649042">
#>           <SimplePredicate field="Income" operator="lessThan" value="35815.64"/>
#>          </Node>
#>          <Node id="8" score="0.159806183888325">
#>           <SimplePredicate field="Income" operator="greaterOrEqual" value="35815.64"/>
#>          </Node>
#>          <Node id="9" score="0.0844573840497526">
#>           <SimplePredicate field="Income" operator="isMissing"/>
#>          </Node>
#>         </Node>
#>         <Node id="10" score="0.0304204715377185">
#>          <SimplePredicate field="Age" operator="isMissing"/>
#>         </Node>
#>        </Node>
#>        <Node id="11" score="0.17957031396641">
#>         <SimplePredicate field="Age" operator="greaterOrEqual" value="40.5"/>
#>        </Node>
#>        <Node id="12" score="0.110626908541706">
#>         <SimplePredicate field="Age" operator="isMissing"/>
#>        </Node>
#>       </Node>
#>       <Node id="13" score="-0.0106683851348229">
#>        <SimplePredicate field="Marital" operator="isMissing"/>
#>       </Node>
#>      </Node>
#>     </TreeModel>
#>    </Segment>
#>    <Segment id="3">
#>     <True/>
#>     <TreeModel modelName="GBM_Model" functionName="regression" algorithmName="gbm" splitCharacteristic="multiSplit">
#>      <MiningSchema>
#>       <MiningField name="Adjusted" usageType="predicted" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Age" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Employment" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Marital" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Income" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Sex" usageType="active" invalidValueTreatment="returnInvalid"/>
#>      </MiningSchema>
#>      <Node id="1">
#>       <True/>
#>       <Node id="2" score="-0.106317273934743">
#>        <SimpleSetPredicate field="Marital" booleanOperator="isIn">
#>         <Array n="5" type="string">&quot;Absent&quot;   &quot;Divorced&quot;   &quot;Married-spouse-absent&quot;   &quot;Unmarried&quot;   &quot;Widowed&quot;</Array>
#>        </SimpleSetPredicate>
#>       </Node>
#>       <Node id="3" score="0.0953529957752467">
#>        <SimpleSetPredicate field="Marital" booleanOperator="isIn">
#>         <Array n="1" type="string">&quot;Married&quot;</Array>
#>        </SimpleSetPredicate>
#>        <Node id="4" score="-0.041578237796967">
#>         <SimplePredicate field="Age" operator="lessThan" value="30.5"/>
#>        </Node>
#>        <Node id="5" score="0.122461209528121">
#>         <SimplePredicate field="Age" operator="greaterOrEqual" value="30.5"/>
#>         <Node id="6" score="0.0926483626779545">
#>          <SimplePredicate field="Income" operator="lessThan" value="49306.67"/>
#>          <Node id="7" score="0.0724919315464392">
#>           <SimpleSetPredicate field="Employment" booleanOperator="isIn">
#>            <Array n="3" type="string">&quot;Consultant&quot;   &quot;Private&quot;   &quot;PSFederal&quot;</Array>
#>           </SimpleSetPredicate>
#>          </Node>
#>          <Node id="8" score="0.182802581920732">
#>           <SimpleSetPredicate field="Employment" booleanOperator="isIn">
#>            <Array n="4" type="string">&quot;PSLocal&quot;   &quot;PSState&quot;   &quot;SelfEmp&quot;   &quot;Volunteer&quot;</Array>
#>           </SimpleSetPredicate>
#>          </Node>
#>          <Node id="9" score="0.0926483626779545">
#>           <SimplePredicate field="Employment" operator="isMissing"/>
#>          </Node>
#>         </Node>
#>         <Node id="10" score="0.218952251484038">
#>          <SimplePredicate field="Income" operator="greaterOrEqual" value="49306.67"/>
#>         </Node>
#>         <Node id="11" score="0.122461209528121">
#>          <SimplePredicate field="Income" operator="isMissing"/>
#>         </Node>
#>        </Node>
#>        <Node id="12" score="0.0953529957752467">
#>         <SimplePredicate field="Age" operator="isMissing"/>
#>        </Node>
#>       </Node>
#>       <Node id="13" score="-0.00601340954789889">
#>        <SimplePredicate field="Marital" operator="isMissing"/>
#>       </Node>
#>      </Node>
#>     </TreeModel>
#>    </Segment>
#>   </Segmentation>
#>  </MiningModel>
#> </PMML>

# Classification example:
mod2 <- gbm(Species ~ .,
  data = iris, n.trees = 2,
  interaction.depth = 3, distribution = "multinomial"
)

# The PMML will include a regression model to read the gbm object outputs
# and convert to a "response" prediction type.
pmml(mod2)

#> Now converting tree  1  to PMML, for category# 1

#> Now converting tree  2  to PMML, for category# 1

#> Now converting tree  1  to PMML, for category# 2

#> Now converting tree  2  to PMML, for category# 2

#> Now converting tree  1  to PMML, for category# 3

#> Now converting tree  2  to PMML, for category# 3

#> <PMML version="4.3" xmlns="http://www.dmg.org/PMML-4_3" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.dmg.org/PMML-4_3 http://www.dmg.org/pmml/v4-3/pmml-4-3.xsd">
#>  <Header copyright="Copyright (c) 2019 dmitriy" description="Generalized Boosted Tree Model">
#>   <Extension name="user" value="dmitriy" extender="SoftwareAG PMML Generator"/>
#>   <Application name="SoftwareAG PMML Generator" version="1.4"/>
#>   <Timestamp>2019-06-04 17:04:22</Timestamp>
#>  </Header>
#>  <DataDictionary numberOfFields="5">
#>   <DataField name="Species" optype="categorical" dataType="string">
#>    <Value value="setosa"/>
#>    <Value value="versicolor"/>
#>    <Value value="virginica"/>
#>   </DataField>
#>   <DataField name="Sepal.Length" optype="continuous" dataType="double"/>
#>   <DataField name="Sepal.Width" optype="continuous" dataType="double"/>
#>   <DataField name="Petal.Length" optype="continuous" dataType="double"/>
#>   <DataField name="Petal.Width" optype="continuous" dataType="double"/>
#>  </DataDictionary>
#>  <MiningModel modelName="GBM_Model" functionName="classification">
#>   <MiningSchema>
#>    <MiningField name="Species" usageType="predicted" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="Sepal.Length" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="Sepal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="Petal.Length" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="Petal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>   </MiningSchema>
#>   <Output>
#>    <OutputField name="Predicted_Species" optype="categorical" dataType="string" feature="predictedValue"/>
#>    <OutputField name="Probability_setosa" optype="continuous" dataType="double" feature="probability" value="setosa"/>
#>    <OutputField name="Probability_versicolor" optype="continuous" dataType="double" feature="probability" value="versicolor"/>
#>    <OutputField name="Probability_virginica" optype="continuous" dataType="double" feature="probability" value="virginica"/>
#>    <OutputField name="link_setosa" optype="continuous" dataType="double" feature="transformedValue" segmentId="2">
#>     <FieldRef field="UpdatedPredictedValue12"/>
#>    </OutputField>
#>    <OutputField name="link_versicolor" optype="continuous" dataType="double" feature="transformedValue" segmentId="4">
#>     <FieldRef field="UpdatedPredictedValue22"/>
#>    </OutputField>
#>    <OutputField name="link_virginica" optype="continuous" dataType="double" feature="transformedValue" segmentId="6">
#>     <FieldRef field="UpdatedPredictedValue32"/>
#>    </OutputField>
#>   </Output>
#>   <Segmentation multipleModelMethod="modelChain">
#>    <Segment id="1">
#>     <True/>
#>     <TreeModel modelName="GBM_Model" functionName="regression" algorithmName="gbm" splitCharacteristic="multiSplit">
#>      <MiningSchema>
#>       <MiningField name="Species" usageType="predicted" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Sepal.Length" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Sepal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Length" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>      </MiningSchema>
#>      <Output>
#>       <OutputField name="UpdatedPredictedValue11" optype="continuous" dataType="double" feature="predictedValue"/>
#>      </Output>
#>      <Node id="1">
#>       <True/>
#>       <Node id="2" score="0.3">
#>        <SimplePredicate field="Petal.Length" operator="lessThan" value="2.35"/>
#>        <Node id="3" score="0.3">
#>         <SimplePredicate field="Sepal.Length" operator="lessThan" value="4.95"/>
#>        </Node>
#>        <Node id="4" score="0.3">
#>         <SimplePredicate field="Sepal.Length" operator="greaterOrEqual" value="4.95"/>
#>        </Node>
#>        <Node id="5" score="0.3">
#>         <SimplePredicate field="Sepal.Length" operator="isMissing"/>
#>        </Node>
#>       </Node>
#>       <Node id="6" score="-0.15">
#>        <SimplePredicate field="Petal.Length" operator="greaterOrEqual" value="2.35"/>
#>        <Node id="7" score="-0.15">
#>         <SimplePredicate field="Petal.Length" operator="lessThan" value="5.45"/>
#>        </Node>
#>        <Node id="8" score="-0.15">
#>         <SimplePredicate field="Petal.Length" operator="greaterOrEqual" value="5.45"/>
#>        </Node>
#>        <Node id="9" score="-0.15">
#>         <SimplePredicate field="Petal.Length" operator="isMissing"/>
#>        </Node>
#>       </Node>
#>       <Node id="10" score="0.018">
#>        <SimplePredicate field="Petal.Length" operator="isMissing"/>
#>       </Node>
#>      </Node>
#>     </TreeModel>
#>    </Segment>
#>    <Segment id="2">
#>     <True/>
#>     <TreeModel modelName="GBM_Model" functionName="regression" algorithmName="gbm" splitCharacteristic="multiSplit">
#>      <MiningSchema>
#>       <MiningField name="Species" usageType="predicted" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Sepal.Length" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Sepal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Length" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="UpdatedPredictedValue11" optype="continuous"/>
#>      </MiningSchema>
#>      <Output>
#>       <OutputField name="TreePredictedValue12" optype="continuous" dataType="double" feature="predictedValue"/>
#>       <OutputField name="UpdatedPredictedValue12" optype="continuous" dataType="double" feature="transformedValue">
#>        <Apply function="+">
#>         <FieldRef field="UpdatedPredictedValue11"/>
#>         <FieldRef field="TreePredictedValue12"/>
#>        </Apply>
#>       </OutputField>
#>      </Output>
#>      <Node id="1">
#>       <True/>
#>       <Node id="2" score="0.227638197604472">
#>        <SimplePredicate field="Petal.Length" operator="lessThan" value="2.45"/>
#>       </Node>
#>       <Node id="3" score="-0.138973320980512">
#>        <SimplePredicate field="Petal.Length" operator="greaterOrEqual" value="2.45"/>
#>        <Node id="4" score="-0.138422961126724">
#>         <SimplePredicate field="Petal.Length" operator="lessThan" value="4.05"/>
#>        </Node>
#>        <Node id="5" score="-0.139151816068227">
#>         <SimplePredicate field="Petal.Length" operator="greaterOrEqual" value="4.05"/>
#>         <Node id="6" score="-0.1394640682052">
#>          <SimplePredicate field="Petal.Length" operator="lessThan" value="5.2"/>
#>         </Node>
#>         <Node id="7" score="-0.13863883041463">
#>          <SimplePredicate field="Petal.Length" operator="greaterOrEqual" value="5.2"/>
#>         </Node>
#>         <Node id="8" score="-0.139151816068227">
#>          <SimplePredicate field="Petal.Length" operator="isMissing"/>
#>         </Node>
#>        </Node>
#>        <Node id="9" score="-0.138973320980512">
#>         <SimplePredicate field="Petal.Length" operator="isMissing"/>
#>        </Node>
#>       </Node>
#>       <Node id="10" score="-0.0118813278710511">
#>        <SimplePredicate field="Petal.Length" operator="isMissing"/>
#>       </Node>
#>      </Node>
#>     </TreeModel>
#>    </Segment>
#>    <Segment id="3">
#>     <True/>
#>     <TreeModel modelName="GBM_Model" functionName="regression" algorithmName="gbm" splitCharacteristic="multiSplit">
#>      <MiningSchema>
#>       <MiningField name="Species" usageType="predicted" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Sepal.Length" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Sepal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Length" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>      </MiningSchema>
#>      <Output>
#>       <OutputField name="UpdatedPredictedValue21" optype="continuous" dataType="double" feature="predictedValue"/>
#>      </Output>
#>      <Node id="1">
#>       <True/>
#>       <Node id="2" score="-0.15">
#>        <SimplePredicate field="Petal.Length" operator="lessThan" value="2.35"/>
#>       </Node>
#>       <Node id="3" score="0.0797872340425532">
#>        <SimplePredicate field="Petal.Length" operator="greaterOrEqual" value="2.35"/>
#>        <Node id="4" score="0.28125">
#>         <SimplePredicate field="Petal.Length" operator="lessThan" value="4.85"/>
#>         <Node id="5" score="0.3">
#>          <SimplePredicate field="Petal.Width" operator="lessThan" value="1.35"/>
#>         </Node>
#>         <Node id="6" score="0.255">
#>          <SimplePredicate field="Petal.Width" operator="greaterOrEqual" value="1.35"/>
#>         </Node>
#>         <Node id="7" score="0.28125">
#>          <SimplePredicate field="Petal.Width" operator="isMissing"/>
#>         </Node>
#>        </Node>
#>        <Node id="8" score="-0.130434782608696">
#>         <SimplePredicate field="Petal.Length" operator="greaterOrEqual" value="4.85"/>
#>        </Node>
#>        <Node id="9" score="0.0797872340425532">
#>         <SimplePredicate field="Petal.Length" operator="isMissing"/>
#>        </Node>
#>       </Node>
#>       <Node id="10" score="-0.00599999999999996">
#>        <SimplePredicate field="Petal.Length" operator="isMissing"/>
#>       </Node>
#>      </Node>
#>     </TreeModel>
#>    </Segment>
#>    <Segment id="4">
#>     <True/>
#>     <TreeModel modelName="GBM_Model" functionName="regression" algorithmName="gbm" splitCharacteristic="multiSplit">
#>      <MiningSchema>
#>       <MiningField name="Species" usageType="predicted" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Sepal.Length" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Sepal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Length" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="UpdatedPredictedValue21" optype="continuous"/>
#>      </MiningSchema>
#>      <Output>
#>       <OutputField name="TreePredictedValue22" optype="continuous" dataType="double" feature="predictedValue"/>
#>       <OutputField name="UpdatedPredictedValue22" optype="continuous" dataType="double" feature="transformedValue">
#>        <Apply function="+">
#>         <FieldRef field="UpdatedPredictedValue21"/>
#>         <FieldRef field="TreePredictedValue22"/>
#>        </Apply>
#>       </OutputField>
#>      </Output>
#>      <Node id="1">
#>       <True/>
#>       <Node id="2" score="-0.138909909815322">
#>        <SimplePredicate field="Petal.Length" operator="lessThan" value="2.45"/>
#>       </Node>
#>       <Node id="3" score="0.0963410279158985">
#>        <SimplePredicate field="Petal.Length" operator="greaterOrEqual" value="2.45"/>
#>        <Node id="4" score="0.23354296951992">
#>         <SimplePredicate field="Petal.Length" operator="lessThan" value="5"/>
#>         <Node id="5" score="0.228921077330639">
#>          <SimplePredicate field="Petal.Width" operator="lessThan" value="1.35"/>
#>         </Node>
#>         <Node id="6" score="0.243248943117411">
#>          <SimplePredicate field="Petal.Width" operator="greaterOrEqual" value="1.35"/>
#>         </Node>
#>         <Node id="7" score="0.23354296951992">
#>          <SimplePredicate field="Petal.Width" operator="isMissing"/>
#>         </Node>
#>        </Node>
#>        <Node id="8" score="-0.139951204846584">
#>         <SimplePredicate field="Petal.Length" operator="greaterOrEqual" value="5"/>
#>        </Node>
#>        <Node id="9" score="0.0963410279158985">
#>         <SimplePredicate field="Petal.Length" operator="isMissing"/>
#>        </Node>
#>       </Node>
#>       <Node id="10" score="0.0147873695024086">
#>        <SimplePredicate field="Petal.Length" operator="isMissing"/>
#>       </Node>
#>      </Node>
#>     </TreeModel>
#>    </Segment>
#>    <Segment id="5">
#>     <True/>
#>     <TreeModel modelName="GBM_Model" functionName="regression" algorithmName="gbm" splitCharacteristic="multiSplit">
#>      <MiningSchema>
#>       <MiningField name="Species" usageType="predicted" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Sepal.Length" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Sepal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Length" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>      </MiningSchema>
#>      <Output>
#>       <OutputField name="UpdatedPredictedValue31" optype="continuous" dataType="double" feature="predictedValue"/>
#>      </Output>
#>      <Node id="1">
#>       <True/>
#>       <Node id="2" score="-0.141346153846154">
#>        <SimplePredicate field="Petal.Length" operator="lessThan" value="4.85"/>
#>        <Node id="3" score="-0.105">
#>         <SimplePredicate field="Sepal.Width" operator="lessThan" value="2.55"/>
#>        </Node>
#>        <Node id="4" score="-0.15">
#>         <SimplePredicate field="Sepal.Width" operator="greaterOrEqual" value="2.55"/>
#>        </Node>
#>        <Node id="5" score="-0.141346153846154">
#>         <SimplePredicate field="Sepal.Width" operator="isMissing"/>
#>        </Node>
#>       </Node>
#>       <Node id="6" score="0.280434782608696">
#>        <SimplePredicate field="Petal.Length" operator="greaterOrEqual" value="4.85"/>
#>        <Node id="7" score="0.255">
#>         <SimplePredicate field="Petal.Length" operator="lessThan" value="5.25"/>
#>        </Node>
#>        <Node id="8" score="0.3">
#>         <SimplePredicate field="Petal.Length" operator="greaterOrEqual" value="5.25"/>
#>        </Node>
#>        <Node id="9" score="0.280434782608696">
#>         <SimplePredicate field="Petal.Length" operator="isMissing"/>
#>        </Node>
#>       </Node>
#>       <Node id="10" score="-0.0120000000000001">
#>        <SimplePredicate field="Petal.Length" operator="isMissing"/>
#>       </Node>
#>      </Node>
#>     </TreeModel>
#>    </Segment>
#>    <Segment id="6">
#>     <True/>
#>     <TreeModel modelName="GBM_Model" functionName="regression" algorithmName="gbm" splitCharacteristic="multiSplit">
#>      <MiningSchema>
#>       <MiningField name="Species" usageType="predicted" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Sepal.Length" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Sepal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Length" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="UpdatedPredictedValue31" optype="continuous"/>
#>      </MiningSchema>
#>      <Output>
#>       <OutputField name="TreePredictedValue32" optype="continuous" dataType="double" feature="predictedValue"/>
#>       <OutputField name="UpdatedPredictedValue32" optype="continuous" dataType="double" feature="transformedValue">
#>        <Apply function="+">
#>         <FieldRef field="UpdatedPredictedValue31"/>
#>         <FieldRef field="TreePredictedValue32"/>
#>        </Apply>
#>       </OutputField>
#>      </Output>
#>      <Node id="1">
#>       <True/>
#>       <Node id="2" score="-0.140203999979924">
#>        <SimplePredicate field="Petal.Length" operator="lessThan" value="5"/>
#>        <Node id="3" score="-0.143832282675387">
#>         <SimplePredicate field="Sepal.Width" operator="lessThan" value="2.55"/>
#>        </Node>
#>        <Node id="4" score="-0.139132007365356">
#>         <SimplePredicate field="Sepal.Width" operator="greaterOrEqual" value="2.55"/>
#>         <Node id="5" score="-0.138936076605078">
#>          <SimplePredicate field="Petal.Length" operator="lessThan" value="4.35"/>
#>         </Node>
#>         <Node id="6" score="-0.139798171950301">
#>          <SimplePredicate field="Petal.Length" operator="greaterOrEqual" value="4.35"/>
#>         </Node>
#>         <Node id="7" score="-0.139132007365356">
#>          <SimplePredicate field="Petal.Length" operator="isMissing"/>
#>         </Node>
#>        </Node>
#>        <Node id="8" score="-0.140203999979924">
#>         <SimplePredicate field="Sepal.Width" operator="isMissing"/>
#>        </Node>
#>       </Node>
#>       <Node id="9" score="0.23009641910618">
#>        <SimplePredicate field="Petal.Length" operator="greaterOrEqual" value="5"/>
#>       </Node>
#>       <Node id="10" score="-0.0513318993992592">
#>        <SimplePredicate field="Petal.Length" operator="isMissing"/>
#>       </Node>
#>      </Node>
#>     </TreeModel>
#>    </Segment>
#>    <Segment id="7">
#>     <True/>
#>     <RegressionModel functionName="classification" normalizationMethod="softmax">
#>      <MiningSchema>
#>       <MiningField name="UpdatedPredictedValue12"/>
#>       <MiningField name="UpdatedPredictedValue22"/>
#>       <MiningField name="UpdatedPredictedValue32"/>
#>      </MiningSchema>
#>      <Output>
#>       <OutputField name="Predicted_Species" optype="categorical" dataType="string" feature="predictedValue"/>
#>       <OutputField name="Probability_setosa" optype="continuous" dataType="double" feature="probability" value="setosa"/>
#>       <OutputField name="Probability_versicolor" optype="continuous" dataType="double" feature="probability" value="versicolor"/>
#>       <OutputField name="Probability_virginica" optype="continuous" dataType="double" feature="probability" value="virginica"/>
#>      </Output>
#>      <RegressionTable intercept="0" targetCategory="setosa">
#>       <NumericPredictor name="UpdatedPredictedValue12" coefficient="1"/>
#>      </RegressionTable>
#>      <RegressionTable intercept="0" targetCategory="versicolor">
#>       <NumericPredictor name="UpdatedPredictedValue22" coefficient="1"/>
#>      </RegressionTable>
#>      <RegressionTable intercept="0" targetCategory="virginica">
#>       <NumericPredictor name="UpdatedPredictedValue32" coefficient="1"/>
#>      </RegressionTable>
#>     </RegressionModel>
#>    </Segment>
#>   </Segmentation>
#>  </MiningModel>
#> </PMML>