Generate PMML for an iForest object from the isofor package.

Generate PMML for an iForest object from the isofor package.

# S3 method for iForest
pmml(model, model_name = "isolationForest_Model",
  app_name = "SoftwareAG PMML Generator",
  description = "Isolation Forest Model", copyright = NULL,
  transforms = NULL, missing_value_replacement = NULL,
  anomalyThreshold = 0.6,
  parent_invalid_value_treatment = "returnInvalid",
  child_invalid_value_treatment = "asIs", ...)

Arguments

model
An iForest object from package isofor.

model_name
A name to be given to the PMML model.

app_name
The name of the application that generated the PMML.

description
A descriptive text for the Header element of the PMML.

copyright
The copyright notice for the model.

transforms
Data transformations.

missing_value_replacement
Value to be used as the 'missingValueReplacement' attribute for all MiningFields.

anomalyThreshold
Double between 0 and 1. Predicted values greater than this are classified as anomalies.

parent_invalid_value_treatment
Invalid value treatment at the top MiningField level.

child_invalid_value_treatment
Invalid value treatment at the model segment MiningField level.

...
Further arguments passed to or from other methods.

Value

PMML representation of the iForest object.

Details

This function converts the iForest model object to the PMML format. The PMML outputs the anomaly score as well as a boolean value indicating whether the input is an anomaly or not. This is done by simply comparing the anomaly score with anomalyThreshold, a parameter in the pmml function. The iForest function automatically adds an extra level to all categorical variables, labelled "."; this is kept in the PMML representation even though the use of this extra factor in the predict function is unclear.

Isolation forest anomaly detection models are not yet supported by DMG PMML schema version 4.3. The PMML produced by this exporter uses an extended schema, and can be consumed by Zementis products. The extended schema is a superset of 4.3, and contains enhancements that will be part of the upcoming PMML 4.4 release.

References

isofor package on GitHub

See also

Examples

# NOT RUN {
# Build iForest model using iris dataset. Create an isolation
# forest with 10 trees. Sample 30 data points at a time from
# the iris dataset to fit the trees.
library(isofor)
data(iris)
mod <- iForest(iris, nt = 10, phi = 30)

# Convert to PMML:
pm <- pmml(mod)
# }