Generate the PMML representation for a kmeans object from the package stats.

The kmeans object (a cluster described by k centroids) is converted into a PMML representation.

# S3 method for kmeans
pmml(model, model_name = "KMeans_Model",
  app_name = "SoftwareAG PMML Generator",
  description = "KMeans cluster model", copyright = NULL,
  transforms = NULL, missing_value_replacement = NULL,
  algorithm.name = "KMeans: Hartigan and Wong", ...)

Arguments

model
A kmeans object.

model_name
A name to be given to the PMML model.

app_name
The name of the application that generated the PMML.

description
A descriptive text for the Header element of the PMML.

copyright
The copyright notice for the model.

transforms
Data transformations.

missing_value_replacement
Value to be used as the 'missingValueReplacement' attribute for all MiningFields.

algorithm.name
The variety of kmeans used.

...
Further arguments passed to or from other methods.

Details

A kmeans object is obtained by applying the kmeans function from the stats package. This method typically requires the user to normalize all the variables; these operations can be done using transforms so that the normalization information is included in PMML.

References

R project: K-Means Clustering

Examples

ds <- rbind(
  matrix(rnorm(100, sd = 0.3), ncol = 2),
  matrix(rnorm(100, mean = 1, sd = 0.3), ncol = 2)
)
colnames(ds) <- c("Dimension1", "Dimension2")
cl <- kmeans(ds, 2)
pmml(cl)

#> <PMML version="4.3" xmlns="http://www.dmg.org/PMML-4_3" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.dmg.org/PMML-4_3 http://www.dmg.org/pmml/v4-3/pmml-4-3.xsd">
#>  <Header copyright="Copyright (c) 2019 dmitriy" description="KMeans cluster model">
#>   <Extension name="user" value="dmitriy" extender="SoftwareAG PMML Generator"/>
#>   <Application name="SoftwareAG PMML Generator" version="1.4"/>
#>   <Timestamp>2019-05-30 15:05:29</Timestamp>
#>  </Header>
#>  <DataDictionary numberOfFields="2">
#>   <DataField name="Dimension1" optype="continuous" dataType="double"/>
#>   <DataField name="Dimension2" optype="continuous" dataType="double"/>
#>  </DataDictionary>
#>  <ClusteringModel modelName="KMeans_Model" functionName="clustering" algorithmName="KMeans: Hartigan and Wong" modelClass="centerBased" numberOfClusters="2">
#>   <MiningSchema>
#>    <MiningField name="Dimension1" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="Dimension2" invalidValueTreatment="returnInvalid"/>
#>   </MiningSchema>
#>   <Output>
#>    <OutputField name="predictedValue" feature="predictedValue" dataType="string" optype="categorical"/>
#>    <OutputField name="clusterAffinity_1" feature="clusterAffinity" value="1" dataType="double"/>
#>    <OutputField name="clusterAffinity_2" feature="clusterAffinity" value="2" dataType="double"/>
#>   </Output>
#>   <ComparisonMeasure kind="distance">
#>    <squaredEuclidean/>
#>   </ComparisonMeasure>
#>   <ClusteringField field="Dimension1" compareFunction="absDiff"/>
#>   <ClusteringField field="Dimension2" compareFunction="absDiff"/>
#>   <Cluster name="1" size="50" id="1">
#>    <Array n="2" type="real">1.04657943542431 1.03059405914148</Array>
#>   </Cluster>
#>   <Cluster name="2" size="50" id="2">
#>    <Array n="2" type="real">-0.0572506663667143 -0.0186524801762376</Array>
#>   </Cluster>
#>  </ClusteringModel>
#> </PMML>