Generate the PMML representation for a naiveBayes object from the package e1071.

Generate the PMML representation for a naiveBayes object from the package e1071.

# S3 method for naiveBayes
pmml(model, model_name = "naiveBayes_Model",
  app_name = "SoftwareAG PMML Generator",
  description = "NaiveBayes Model", copyright = NULL,
  transforms = NULL, missing_value_replacement = NULL, predicted_field,
  ...)

Arguments

model
A naiveBayes object.

model_name
A name to be given to the PMML model.

app_name
The name of the application that generated the PMML.

description
A descriptive text for the Header element of the PMML.

copyright
The copyright notice for the model.

transforms
Data transformations.

missing_value_replacement
Value to be used as the 'missingValueReplacement' attribute for all MiningFields.

predicted_field
Required parameter; the name of the predicted field.

...
Further arguments passed to or from other methods.

Value

PMML representation of the naiveBayes object.

Details

The PMML representation of the NaiveBayes model implements the definition as specified by the Data Mining Group: intermediate probability values which are less than the threshold value are replaced by the threshold value. This is different from the prediction function of the e1071 in which only probability values of 0 and standard deviations of continuous variables of with the value 0 are replaced by the threshold value. The two values will therefore not match exactly for cases involving very small probabilities.

References

e1071: Misc Functions of the Department of Statistics, Probability Theory Group (Formerly: E1071), TU Wien (on CRAN)

A. Guazzelli, T. Jena, W. Lin, M. Zeller (2013). Extending the Naive Bayes Model Element in PMML: Adding Support for Continuous Input Variables. In Proceedings of the 19th ACM SIGKDD Conference on Knowledge Discovery and Data Mining.

Examples

library(e1071)

data(houseVotes84)
house <- na.omit(houseVotes84)

model <- naiveBayes(Class ~ V1 + V2 + V3, data = house, threshold = 0.003)

pmml(model, dataset = house, predicted_field = "Class")

#> <PMML version="4.3" xmlns="http://www.dmg.org/PMML-4_3" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.dmg.org/PMML-4_3 http://www.dmg.org/pmml/v4-3/pmml-4-3.xsd">
#>  <Header copyright="Copyright (c) 2019 dmitriy" description="NaiveBayes Model">
#>   <Extension name="user" value="dmitriy" extender="SoftwareAG PMML Generator"/>
#>   <Application name="SoftwareAG PMML Generator" version="1.4"/>
#>   <Timestamp>2019-06-02 13:32:28</Timestamp>
#>  </Header>
#>  <DataDictionary numberOfFields="4">
#>   <DataField name="Class" optype="categorical" dataType="string">
#>    <Value value="democrat"/>
#>    <Value value="republican"/>
#>   </DataField>
#>   <DataField name="V1" optype="categorical" dataType="string">
#>    <Value value="n"/>
#>    <Value value="y"/>
#>   </DataField>
#>   <DataField name="V2" optype="categorical" dataType="string">
#>    <Value value="n"/>
#>    <Value value="y"/>
#>   </DataField>
#>   <DataField name="V3" optype="categorical" dataType="string">
#>    <Value value="n"/>
#>    <Value value="y"/>
#>   </DataField>
#>  </DataDictionary>
#>  <NaiveBayesModel modelName="naiveBayes_Model" functionName="classification" threshold="0.003">
#>   <MiningSchema>
#>    <MiningField name="Class" usageType="predicted" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="V1" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="V2" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="V3" usageType="active" invalidValueTreatment="returnInvalid"/>
#>   </MiningSchema>
#>   <Output>
#>    <OutputField name="Predicted_Class" optype="categorical" dataType="string" feature="predictedValue"/>
#>    <OutputField name="Probability_democrat" optype="continuous" dataType="double" feature="probability" value="democrat"/>
#>    <OutputField name="Probability_republican" optype="continuous" dataType="double" feature="probability" value="republican"/>
#>   </Output>
#>   <BayesInputs>
#>    <BayesInput fieldName="V1">
#>     <PairCounts value="n">
#>      <TargetValueCounts>
#>       <TargetValueCount value="democrat" count="51"/>
#>       <TargetValueCount value="republican" count="85"/>
#>      </TargetValueCounts>
#>     </PairCounts>
#>     <PairCounts value="y">
#>      <TargetValueCounts>
#>       <TargetValueCount value="democrat" count="73"/>
#>       <TargetValueCount value="republican" count="23"/>
#>      </TargetValueCounts>
#>     </PairCounts>
#>    </BayesInput>
#>    <BayesInput fieldName="V2">
#>     <PairCounts value="n">
#>      <TargetValueCounts>
#>       <TargetValueCount value="democrat" count="68"/>
#>       <TargetValueCount value="republican" count="57"/>
#>      </TargetValueCounts>
#>     </PairCounts>
#>     <PairCounts value="y">
#>      <TargetValueCounts>
#>       <TargetValueCount value="democrat" count="56"/>
#>       <TargetValueCount value="republican" count="51"/>
#>      </TargetValueCounts>
#>     </PairCounts>
#>    </BayesInput>
#>    <BayesInput fieldName="V3">
#>     <PairCounts value="n">
#>      <TargetValueCounts>
#>       <TargetValueCount value="democrat" count="18"/>
#>       <TargetValueCount value="republican" count="91"/>
#>      </TargetValueCounts>
#>     </PairCounts>
#>     <PairCounts value="y">
#>      <TargetValueCounts>
#>       <TargetValueCount value="democrat" count="106"/>
#>       <TargetValueCount value="republican" count="17"/>
#>      </TargetValueCounts>
#>     </PairCounts>
#>    </BayesInput>
#>   </BayesInputs>
#>   <BayesOutput fieldName="Class">
#>    <TargetValueCounts>
#>     <TargetValueCount value="democrat" count="124"/>
#>     <TargetValueCount value="republican" count="108"/>
#>    </TargetValueCounts>
#>   </BayesOutput>
#>  </NaiveBayesModel>
#> </PMML>