Generate the PMML representation for a nnet object from package nnet.

Generate the PMML representation for a nnet object from package nnet.

# S3 method for nnet
pmml(model, model_name = "NeuralNet_model",
  app_name = "SoftwareAG PMML Generator",
  description = "Neural Network Model", copyright = NULL,
  transforms = NULL, missing_value_replacement = NULL, ...)

Arguments

model
A nnet object.

model_name
A name to be given to the PMML model.

app_name
The name of the application that generated the PMML.

description
A descriptive text for the Header element of the PMML.

copyright
The copyright notice for the model.

transforms
Data transformations.

missing_value_replacement
Value to be used as the 'missingValueReplacement' attribute for all MiningFields.

...
Further arguments passed to or from other methods.

Value

PMML representation of the nnet object.

Details

This function supports both regression and classification neural network models. The model is represented in the PMML NeuralNetwork format.

References

nnet: Feed-forward Neural Networks and Multinomial Log-Linear Models (on CRAN)

Examples

library(nnet)
fit <- nnet(Species ~ ., data = iris, size = 4)

#> # weights:  35
#> initial  value 168.526082 
#> iter  10 value 66.525796
#> iter  20 value 41.928309
#> iter  30 value 7.472947
#> iter  40 value 6.153270
#> iter  50 value 5.461043
#> iter  60 value 3.645428
#> iter  70 value 3.528160
#> iter  80 value 3.268686
#> iter  90 value 3.182114
#> iter 100 value 3.065146
#> final  value 3.065146 
#> stopped after 100 iterations

pmml(fit)

#> <PMML version="4.3" xmlns="http://www.dmg.org/PMML-4_3" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.dmg.org/PMML-4_3 http://www.dmg.org/pmml/v4-3/pmml-4-3.xsd">
#>  <Header copyright="Copyright (c) 2019 dmitriy" description="Neural Network Model">
#>   <Extension name="user" value="dmitriy" extender="SoftwareAG PMML Generator"/>
#>   <Application name="SoftwareAG PMML Generator" version="1.4"/>
#>   <Timestamp>2019-06-04 17:04:26</Timestamp>
#>  </Header>
#>  <DataDictionary numberOfFields="5">
#>   <DataField name="Species" optype="categorical" dataType="string">
#>    <Value value="setosa"/>
#>    <Value value="versicolor"/>
#>    <Value value="virginica"/>
#>   </DataField>
#>   <DataField name="Sepal.Length" optype="continuous" dataType="double"/>
#>   <DataField name="Sepal.Width" optype="continuous" dataType="double"/>
#>   <DataField name="Petal.Length" optype="continuous" dataType="double"/>
#>   <DataField name="Petal.Width" optype="continuous" dataType="double"/>
#>  </DataDictionary>
#>  <NeuralNetwork modelName="NeuralNet_model" functionName="classification" numberOfLayers="2" activationFunction="logistic">
#>   <MiningSchema>
#>    <MiningField name="Species" usageType="predicted" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="Sepal.Length" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="Sepal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="Petal.Length" usageType="active" invalidValueTreatment="returnInvalid"/>
#>    <MiningField name="Petal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>   </MiningSchema>
#>   <Output>
#>    <OutputField name="Predicted_Species" optype="categorical" dataType="string" feature="predictedValue"/>
#>    <OutputField name="Probability_setosa" optype="continuous" dataType="double" feature="probability" value="setosa"/>
#>    <OutputField name="Probability_versicolor" optype="continuous" dataType="double" feature="probability" value="versicolor"/>
#>    <OutputField name="Probability_virginica" optype="continuous" dataType="double" feature="probability" value="virginica"/>
#>   </Output>
#>   <NeuralInputs numberOfInputs="4">
#>    <NeuralInput id="1">
#>     <DerivedField name="derivedNI_Sepal.Length" optype="continuous" dataType="double">
#>      <FieldRef field="Sepal.Length"/>
#>     </DerivedField>
#>    </NeuralInput>
#>    <NeuralInput id="2">
#>     <DerivedField name="derivedNI_Sepal.Width" optype="continuous" dataType="double">
#>      <FieldRef field="Sepal.Width"/>
#>     </DerivedField>
#>    </NeuralInput>
#>    <NeuralInput id="3">
#>     <DerivedField name="derivedNI_Petal.Length" optype="continuous" dataType="double">
#>      <FieldRef field="Petal.Length"/>
#>     </DerivedField>
#>    </NeuralInput>
#>    <NeuralInput id="4">
#>     <DerivedField name="derivedNI_Petal.Width" optype="continuous" dataType="double">
#>      <FieldRef field="Petal.Width"/>
#>     </DerivedField>
#>    </NeuralInput>
#>   </NeuralInputs>
#>   <NeuralLayer numberOfNeurons="4">
#>    <Neuron id="5" bias="-0.800249854802091">
#>     <Con from="1" weight="-2.47317538942221"/>
#>     <Con from="2" weight="-0.823454056603268"/>
#>     <Con from="3" weight="-1.40302481449546"/>
#>     <Con from="4" weight="-0.692062677061383"/>
#>    </Neuron>
#>    <Neuron id="6" bias="96.7255269230573">
#>     <Con from="1" weight="-35.1234878244522"/>
#>     <Con from="2" weight="-34.1867289567867"/>
#>     <Con from="3" weight="97.8719909905898"/>
#>     <Con from="4" weight="-178.247543043077"/>
#>    </Neuron>
#>    <Neuron id="7" bias="19.5518052115713">
#>     <Con from="1" weight="13.1979674409845"/>
#>     <Con from="2" weight="-99.6511994239316"/>
#>     <Con from="3" weight="43.8539703185835"/>
#>     <Con from="4" weight="71.2227412639309"/>
#>    </Neuron>
#>    <Neuron id="8" bias="378.059809599583">
#>     <Con from="1" weight="30.1884976409368"/>
#>     <Con from="2" weight="23.3481615177091"/>
#>     <Con from="3" weight="-114.308559920023"/>
#>     <Con from="4" weight="-46.1353746249672"/>
#>    </Neuron>
#>   </NeuralLayer>
#>   <NeuralLayer numberOfNeurons="3" activationFunction="identity" normalizationMethod="softmax">
#>    <Neuron id="9" bias="44.2309767123624">
#>     <Con from="5" weight="0.13548063956016"/>
#>     <Con from="6" weight="-4.56207720707846"/>
#>     <Con from="7" weight="-287.82612587277"/>
#>     <Con from="8" weight="152.757429718426"/>
#>    </Neuron>
#>    <Neuron id="10" bias="-78.9555226886006">
#>     <Con from="5" weight="0.486628787947548"/>
#>     <Con from="6" weight="50.9927993219285"/>
#>     <Con from="7" weight="151.125595289569"/>
#>     <Con from="8" weight="102.881514219698"/>
#>    </Neuron>
#>    <Neuron id="11" bias="34.0578357084074">
#>     <Con from="5" weight="-0.38510085290272"/>
#>     <Con from="6" weight="-46.0387967752994"/>
#>     <Con from="7" weight="137.0928161561"/>
#>     <Con from="8" weight="-255.26572590547"/>
#>    </Neuron>
#>   </NeuralLayer>
#>   <NeuralOutputs numberOfOutputs="3">
#>    <NeuralOutput outputNeuron="9">
#>     <DerivedField name="derivedNO_Species" optype="continuous" dataType="double">
#>      <NormDiscrete field="Species" value="setosa"/>
#>     </DerivedField>
#>    </NeuralOutput>
#>    <NeuralOutput outputNeuron="10">
#>     <DerivedField name="derivedNO_Species" optype="continuous" dataType="double">
#>      <NormDiscrete field="Species" value="versicolor"/>
#>     </DerivedField>
#>    </NeuralOutput>
#>    <NeuralOutput outputNeuron="11">
#>     <DerivedField name="derivedNO_Species" optype="continuous" dataType="double">
#>      <NormDiscrete field="Species" value="virginica"/>
#>     </DerivedField>
#>    </NeuralOutput>
#>   </NeuralOutputs>
#>  </NeuralNetwork>
#> </PMML>

rm(fit)