Generate the PMML representation for an rfsrc object from the package randomForestSRC.

Generate the PMML representation for an rfsrc object from the package randomForestSRC.

# S3 method for rfsrc
pmml(
  model,
  model_name = "rsf_Model",
  app_name = "SoftwareAG PMML Generator",
  description = "Random Survival Forest Model",
  copyright = NULL,
  transforms = NULL,
  missing_value_replacement = NULL,
  ...
)

Arguments

model
An rfsrc object returned by the function rfsrc with the parameter forest=TRUE.

model_name
A name to be given to the PMML model.

app_name
The name of the application that generated the PMML.

description
A descriptive text for the Header element of the PMML.

copyright
The copyright notice for the model.

transforms
Data transformations.

missing_value_replacement
Value to be used as the 'missingValueReplacement' attribute for all MiningFields.

...
Further arguments passed to or from other methods.

Value

PMML representation of the rfsrc object.

Details

This function is used to export the geometry of the forest to other PMML compliant applications, including graphics packages that are capable of printing binary trees.

The pmml package supports randomSurvivalForest up to version 2.5.0.

References

H. Ishwaran, U.B. Kogalur, E.H. Blackstone, M.S. Lauer (2008), RANDOM SURVIVAL FORESTS. The Annals of Applied Statistics, Vol. 2, No. 3, 841-860.

H. Ishwaran and Udaya B. Kogalur (2006). Random Survival Forests. Cleveland Clinic Technical Report.

Examples

if (FALSE) {
library(randomForestSRC)

data(veteran)

veteran_mod <- rfsrc(Surv(time, status) ~ .,
  data = veteran,
  ntree = 5, forest = TRUE,
  membership = TRUE
)

veteran_mod_pmml <- pmml(veteran_mod)
}